driver-surcharge-experiments-calc/requirements.txt
git+https://gitlab+deploy-token-83:UVXbZBQUGhLdTgxsxtxy@gitlab.city-srv.ru/public_city/citymobil-python-mysql-wrapper.git@0.2.0#egg=citymobil_python_mysql_wrapper
python-dotenv==0.14.0
numpy==1.20.1
pandas==1.2.2
patsy==0.5.1
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.0
six==1.15.0
statsmodels==0.12.2
